requirements.txt
boto3==1.33.11
botocore==1.33.11
celery==5.3.6
Django==5.0
django-filter==23.5
drf-yasg==1.21.7
djangorestframework==3.14.0
numpy==1.26.2
pandas==2.1.4
Pillow==10.1.0
psycopg2-binary==2.9.9
requests==2.31.0

django-elasticsearch-dsl==8.0
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
